Implemented the GridMap. Fixed the use of the MedianFilter in the robot program

+++ KartierungsServer/src/de/dhbw_stuttgart/hb/inf2016/RaumKartierung/Server/Controlling/Controlling.java
@@ -32,7 +32,7 @@ public class Controlling {
 		
 		 if(doesItFit){
 			//If the robot can perform a forward move he will perform it.
-			NextMove = new Forward((int)config.getConstbyName("distancePerMove"),config);
+			NextMove = new Forward((double)config.getConstbyName("distancePerMove"),config);
             return NextMove;
 		}
 		else{
